+++ Ballgorithm_Web/HashCache.hpp
@@ -5,11 +5,13 @@
 # include <optional>
 # include <concepts>
 
+/*
 // T が std::hash を持つかをチェックするコンセプト
 template <typename T>
 concept is_std_hashable = requires(const T& val) {
-	{ std::hash<T>{}(val) } -> std::same_as<std::size_t>;
+	{ std::hash<T>{}(val) } -> std::same_as<std::int32>;
 };
+*/
 
 // ハッシュ値を結合して作るヘルパ
 template <class... Args>
@@ -41,13 +43,13 @@ public:
 	const T* operator->() const { return &_value; }
 
 	// ハッシュ値を計算または取得（イテラブル型は各要素のハッシュを組み合わせる）
-	std::size_t hash() const {
+	std::int32 hash() const {
 		if (!_hash.has_value()) {
 			if constexpr (is_std_hashable<T>) {
 				_hash = std::hash<T>{}(_value);
 			}
 			else if constexpr (std::ranges::range<T> && is_std_hashable<std::ranges::range_value_t<T>>) {
-				std::size_t seed = 0;
+				std::int32 seed = 0;
 				for (const auto& item : _value) {
 					Hash::Combine(seed, item);
 				}
@@ -70,14 +72,14 @@ public:
 
 private:
 	T _value{};
-	mutable std::optional<std::size_t> _hash;
+	mutable std::optional<std::int32> _hash;
 };
 
 // HashCache のための std::hash 特殊化
 namespace std {
 	template <typename T>
 	struct hash<HashCache<T>> {
-		std::size_t operator()(const HashCache<T>& hc) const noexcept {
+		std::int32 operator()(const HashCache<T>& hc) const noexcept {
 			return hc.hash();
 		}
 	};
